+++ packages/gamut/src/GridForm/GridFormInputGroup/index.tsx
@@ -20,7 +20,6 @@ import { GridFormTextArea } from './GridFormTextArea';
 import { GridFormTextInput } from './GridFormTextInput';
 
 export type GridFormInputGroupProps = {
-  clearErrors: UseFormMethods['clearErrors'];
   error?: string;
   isFirstError?: boolean;
   field: GridFormField;
@@ -31,7 +30,6 @@ export type GridFormInputGroupProps = {
 };
 
 export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
-  clearErrors,
   error,
   isFirstError,
   field,
@@ -48,7 +46,6 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
       case 'checkbox':
         return (
           <GridFormCheckboxInput
-            clearErrors={clearErrors}
             field={field}
             register={register}
             showRequired={isRequired}
@@ -59,7 +56,6 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
       case 'custom-group':
         return (
           <GridFormCustomInput
-            clearErrors={clearErrors}
             field={field}
             register={register}
             setValue={setValue}
@@ -70,7 +66,6 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
       case 'radio-group':
         return (
           <GridFormRadioGroupInput
-            clearErrors={clearErrors}
             field={field}
             register={register}
             showRequired={isRequired}
@@ -81,7 +76,6 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
       case 'select':
         return (
           <GridFormSelectInput
-            clearErrors={clearErrors}
             error={!!errorMessage}
             field={field}
             register={register}
@@ -92,7 +86,6 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
       case 'file':
         return (
           <GridFormFileInput
-            clearErrors={clearErrors}
             error={!!errorMessage}
             field={field}
             register={register}
@@ -103,7 +96,6 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
       case 'textarea':
         return (
           <GridFormTextArea
-            clearErrors={clearErrors}
             error={!!errorMessage}
             field={field}
             register={register}
@@ -125,7 +117,6 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
       default:
         return (
           <GridFormTextInput
-            clearErrors={clearErrors}
             error={!!errorMessage}
             field={field}
             register={register}