+++ packages/gamut/src/GridForm/index.tsx
@@ -82,7 +82,6 @@ export function GridForm<
   showRequired = false,
 }: GridFormProps<Values>) {
   const {
-    clearErrors,
     errors,
     handleSubmit,
     register,
@@ -120,7 +119,6 @@ export function GridForm<
 
           return (
             <GridFormInputGroup
-              clearErrors={clearErrors}
               error={errorMessage as string}
               isFirstError={isFirstError}
               field={field}